Try splitting the code a different way
