+++ packages/studio-base/src/components/PanelCatalog/PanelListItem.tsx
