+++ packages/studio-base/src/components/PanelCatalog/index.ts
@@ -2,4 +2,4 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-export * from "./PanelList";
+export * from "./PanelCatalog";
