--- packages/studio-base/src/components/PanelGrid/PanelGrid.stories.tsx
@@ -1,138 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/
-//
-// This file incorporates work covered by the following copyright and
-// permission notice:
-//
-//   Copyright 2018-2021 Cruise LLC
-//
-//   This source code is licensed under the Apache License, Version 2.0,
-//   found at http://www.apache.org/licenses/LICENSE-2.0
-//   You may not use this file except in compliance with the License.
-
-import { useTheme } from "@mui/material";
-import { Meta, StoryFn, StoryObj } from "@storybook/react";
-import { userEvent } from "@storybook/testing-library";
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
-
-import Panel from "@foxglove/studio-base/components/Panel";
-import { PanelGrid } from "@foxglove/studio-base/components/PanelGrid/PanelGrid";
-import PanelCatalogContext, {
-  PanelCatalog,
-  PanelInfo,
-} from "@foxglove/studio-base/context/PanelCatalogContext";
-import MockCurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLayoutProvider/MockCurrentLayoutProvider";
-
-const SamplePanel1 = () => <div />;
-SamplePanel1.panelType = "sample";
-SamplePanel1.defaultConfig = {};
-
-const SamplePanel2 = () => <div />;
-SamplePanel2.panelType = "sample2";
-SamplePanel2.defaultConfig = {};
-
-const MockPanel1 = Panel(SamplePanel1);
-const MockPanel2 = Panel(SamplePanel2);
-
-const allPanels: PanelInfo[] = [
-  { title: "Regular Panel BBB", type: "Sample1", module: async () => ({ default: MockPanel1 }) },
-  { title: "Regular Panel AAA", type: "Sample2", module: async () => ({ default: MockPanel2 }) },
-  {
-    title: "Preconfigured Panel AAA",
-    type: "Sample1",
-    description: "Panel description",
-    module: async () => ({ default: MockPanel1 }),
-    config: { text: "def" },
-  },
-  {
-    title: "Preconfigured Panel BBB",
-    type: "Sample2",
-    module: async () => ({ default: MockPanel1 }),
-    config: { num: 456 },
-  },
-];
-
-class MockPanelCatalog implements PanelCatalog {
-  public getPanels(): readonly PanelInfo[] {
-    return allPanels;
-  }
-  public getPanelByType(type: string): PanelInfo | undefined {
-    return allPanels.find((panel) => !panel.config && panel.type === type);
-  }
-}
-
-type Args = {
-  inputValue?: string;
-  events?: string[];
-};
-
-export default {
-  title: "components/PanelGrid",
-  component: () => <PanelGrid onPanelSelect={() => {}} />,
-  args: {},
-  parameters: { colorScheme: "dark" },
-  decorators: [
-    (Wrapped: StoryFn): JSX.Element => {
-      const theme = useTheme();
-      return (
-        <DndProvider backend={HTML5Backend}>
-          <PanelCatalogContext.Provider value={new MockPanelCatalog()}>
-            <MockCurrentLayoutProvider>
-              <div
-                style={{ margin: 50, height: 480, backgroundColor: theme.palette.background.paper }}
-              >
-                <Wrapped />
-              </div>
-            </MockCurrentLayoutProvider>
-          </PanelCatalogContext.Provider>
-        </DndProvider>
-      );
-    },
-  ],
-  play: async ({ args }) => {
-    userEvent.keyboard(args.inputValue!);
-    args.events?.map((keypress) => userEvent.keyboard(keypress));
-  },
-} as Meta<Args>;
-
-type Story = StoryObj<Args>;
-
-export const Default: Story = {};
-
-export const Filtered: Story = {
-  args: { inputValue: "AAA" },
-};
-
-export const FilteredByDescription: Story = {
-  args: { inputValue: "description" },
-};
-
-export const NoResults: Story = {
-  args: { inputValue: "WWW" },
-  name: "Filtered panel grid without results in any category",
-};
-
-export const CaseInsensitiveFilter: Story = {
-  args: { inputValue: "pA" },
-  name: "Case-insensitive filtering and highlight submenu",
-};
-
-export const Chinese: Story = {
-  parameters: { forceLanguage: "zh" },
-};
-
-export const Japanese: Story = {
-  parameters: { forceLanguage: "ja" },
-};
-
-export const NoResultsChinese: Story = {
-  ...NoResults,
-  parameters: { forceLanguage: "zh" },
-};
-
-export const NoResultsJapanese: Story = {
-  ...NoResults,
-  parameters: { forceLanguage: "ja" },
-};
